LayoutTests/fast/forms/ValidityState-tooLong-input-expected.txt

Tests for tooLong flag with <input> elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.validity.tooLong is false
PASS input.value.length is 5
PASS input.validity.tooLong is false
PASS input.value.length is 6
PASS input.validity.tooLong is false
PASS input.value.length is 4
PASS input.validity.tooLong is true
PASS input.validity.tooLong is false
PASS input.validity.tooLong is true
PASS input.validity.tooLong is true
PASS input.validity.tooLong is false
PASS successfullyParsed is true

TEST COMPLETE
